+++ scripts/mongodb-scripts/backup-remote.sh
@@ -1,3 +1,7 @@
+#!/bin/bash
+
+# Download gzipped collections from production server.
+
 if [ "$#" -ne 3 ]; then
   echo "Usage: $0 <SSH_REMOTE> <SSH_USERNAME> <REMOTE_OPENWHYD_DIR>" >&2
   exit 1